Changed the size of the index of defmethods to be unsigned instead of shorts.
I actually overflowed the method count quite badly when it was a short.

+++ genrcfun.h
@@ -72,7 +72,7 @@ struct restriction
 
 struct method
   {
-   short index;
+   unsigned index;
    unsigned busy;
    short restrictionCount;
    short minRestrictions;
